backend/requirements.txt
Django==4.2.5
djangorestframework==3.15.2
scrypt==0.9.4
selenium==4.14.0
PyJWT==2.7.0
django-cors-headers==3.7.0
drf-spectacular[sidecar]==0.27.2
coverage==6.4.4
GitPython==3.1.27
gunicorn==20.1.0
ruff==0.12.9